% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEA.R
\name{CEA}
\alias{CEA}
\title{CEA}
\usage{
CEA(resultmap, active_gene, d = 0, times = 1, trace = TRUE)
}
\arguments{
\item{resultmap}{The annotation matrix of enrichment analysis. Rows are genes and columns are categories.}

\item{active_gene}{The input character active gene list.}

\item{d}{A stochastic tolerance paramter to add randomness to CEA algorithm.}

\item{times}{The repeated time of CEA algorithm. Positive integer variable.}

\item{trace}{Logical variable indicated whether tracing information on the solving progress is produced.}
}
\value{
This function will return a list with the following components:
  \item{p.values}{The sorted Fisher's exact test p-values of the identified category sets.}
  \item{coverage}{The related coverage of the corresponding category set.}
  \item{category}{The identified category set corresponding to the sorted p-values.}
  \item{annotation}{The pre-processed original annotation matrix.}
}
\description{
A novel combination-based method for gene set functional enrichment analysis.
}
\details{
This function performs the CEA algorithm to perform the enrichment analysis.
CEA is based on a multi-objective optimization problem, 
and the adapted IMPROVED GREEDY algorithm was used to approximatively solve the problem.
}
\references{
Duan-Chen Sun, Yin-Liang Liu, Miao Li, Xiang-Sun Zhang, Ling-Yun Wu.
CEA: Combination-based gene set functional enrichment analysis.
Scientific reports, under review, 2018.
}
