% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ceat.R
\name{ceat}
\alias{ceat}
\title{Combination enrichment analysis tool}
\usage{
ceat(core.sets, gene.set, method = "LP1")
}
\description{
Functional combination enrichment anslysis of gene sets
}
\details{
This is a novel functional enrichment analysis tool based on combined functional terms.
}
\seealso{
\code{\link{get_core_sets}}
}

