% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopTea-package.R
\docType{package}
\name{CopTea-package}
\alias{CopTea}
\alias{CopTea-package}
\title{CopTea - Combinatorial optimization based enrichment analysis}
\description{
This package provides several combinatorial optimization based enrichment analysis
tools for biomedical research.
}
\details{
This package implemented a collection of functional enrichment analysis methods:

\itemize{
  \item \code{\link{cea}} Combinatorial enrichment analysis
  \item \code{\link{netgen}} Network enhanced generative model for enrichment analysis
  \item \code{\link{neeat}} Network enhanced enrichment analysis
  \item \code{\link{noa}} Network ontology analysis
}
}
\references{
Ling-Yun Wu. Combinatorial optimization based enrichment analysis tool. Manuscript,
2016.
}
\keyword{package}

