% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_annotated_genes}
\alias{get_annotated_genes}
\title{Get annotated gene list}
\usage{
get_annotated_genes(annotations, species = "Human", gene.symbol = TRUE)
}
\arguments{
\item{annotations}{A logical matrix indicated the annotations, with each row represents a gene and each column denotes a term.
The row names and column names are set as the corresponding gene and term ID respectively.
See \code{\link{get_annotation}} for more information.}

\item{species}{A string indicated the species name used for translating gene ID to symbols.}

\item{gene.symbol}{A logical variable indicated whether use the gene symbol in the output instead of the central ID.}
}
\value{
This function will return a vector of strings, which is the concatenated IDs (or symbols) of the genes annotated
by the corresponding term.
}
\description{
Extract the genes associated with terms from the annotation matrix.
}
\seealso{
\code{\link{get_annotation}}
}

