% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_annotations}
\alias{get_annotations}
\title{Get annotations from databases}
\usage{
get_annotations(species, STRING.version = "9_1", STRING.threshold = 900)
}
\arguments{
\item{species}{A string indicated species name.}

\item{STRING.version}{A string indicated the version number of STRING database used for
extracting protein interaction network.}

\item{STRING.threshold}{A integer number (0-1000) indicated the threshold for extracting
network from STRING database.}
}
\value{
This function will return a list with the following components:
  \item{annotations}{A logical matrix indicated the annotations, with each row represents
  a gene and each column denotes a term. The row names and column names are set as the
  corresponding gene and term ID respectively.}
  \item{network}{A logical matrix indicated the interaction between genes, which is aggregated
  from the protein interaction network.}
  \item{gene2protein}{A logical matrix indicated the mapping from gene to protein, with
  each row represents a gene and each column denotes a protein. The row names and column
  names are set as the corresponding gene and protein ID respectively.}
  \item{genes}{A vector of gene IDs.}
  \item{terms}{A vector of term IDs.}
  \item{proteins}{A vector of protein IDs.}
  \item{term.info}{A matrix of term information, with three columns: Term ID, Term category,
  and Term description. The row names are set as Term ID.}
}
\description{
Get the annotations needed for functional enrichment analysis from the databases, including GO
and KEGG pathway annotations, gene to protein mapping, protein interaction network.
}

