% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene_sets}
\alias{get_gene_sets}
\title{Transform the gene lists into the matrix of gene sets}
\usage{
get_gene_sets(gene.lists, all.gene)
}
\arguments{
\item{gene.lists}{A list of vectors of gene ids.}

\item{all.gene}{A vector of all gene ids.}
}
\value{
This function returns a sparse matrix as required by enrichment analysis functions.
}
\description{
Generate the gene sets matrix from several gene lists.
}
\details{
This function generates the \code{gene.sets} matrix required by enrichment analysis functions.
The name of each gene set is given by the name of corresponding component of \code{gene.lists},
and assigned to the column of \code{gene.sets}.
}
