% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_GO_annotations}
\alias{get_GO_annotations}
\title{Extract GO temrs information from GO annotation}
\usage{
get_GO_annotations(go.map, evidence = "ALL", category = "ALL",
  gene.set = NULL)
}
\arguments{
\item{go.map}{An object of class "Go3AnnDbBimap".}

\item{evidence}{Vector of string to filter the GO annotations, could be "ALL" or a set of evidence codes.}

\item{category}{Vector of string to filter the GO categories, could be "ALL" or a set of GO categories.}

\item{gene.set}{Vector of string to filter the genes.}
}
\value{
This function returns a sparse matrix as required by enrichment analysis methods,
with each row represents a gene and each column denotes a term.
The row names and column names are set as the corresponding gene and term ID respectively.
}
\description{
Generate the annotation matrix from an object of class "Go3AnnDbBimap" in Bioconductor annotation package.
}
\details{
This function generates the annotation matrix required by enrichment analysis methods.
}
\examples{

\dontrun{
source("http://bioconductor.org/biocLite.R")
biocLite("org.Hs.eg.db")
library(org.Hs.eg.db)
x <- get_GO_annotations(org.Hs.egGO2ALLEGS)
}

}
\seealso{
\code{\link{cea}}, \code{\link{netgen}}
}

