% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapping.R
\name{id_mapping}
\alias{id_mapping}
\title{Mapping gene symbols to central id}
\usage{
id_mapping(symbols, species = "Human", warning.unmapped = TRUE)
}
\description{
Mapping gene symbols to central id
}

