% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{id_mapping_from_symbol}
\alias{id_mapping_from_symbol}
\title{Mapping gene symbols to central id}
\usage{
id_mapping_from_symbol(symbols, species = "Human", warning.unmapped = TRUE)
}
\description{
Mapping gene symbols to central id
}

