% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapping.R
\name{id_mapping_species}
\alias{id_mapping_species}
\title{Mapping gene central id from one species to other species}
\usage{
id_mapping_species(ids, species.from, species.to, rm.na = FALSE)
}
\description{
Mapping gene central id from one species to other species
}

