% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{id_mapping_to_symbol}
\alias{id_mapping_to_symbol}
\title{Mapping central id to gene symbols}
\usage{
id_mapping_to_symbol(ids, species = "Human", warning.unmapped = TRUE)
}
\description{
Mapping central id to gene symbols
}

