% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{NEEAT-package}
\alias{NEEAT}
\alias{NEEAT-package}
\title{NEEAT - Network enhanced enrichment analysis tools}
\description{
This package provides the network enhanced enrichment analysis tool for biomedical research.
}
\details{
This package implemented a collection of functional enrichment analysis methods:

\itemize{
  \item \code{\link{neeat}} Network enhanced enrichment analysis tool
  \item \code{\link{noa}} Network ontology analysis
}
}
\references{
Ling-Yun Wu. Network enhanced enrichment analysis tool. Manuscript,
2014.
}
\keyword{package}

