% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neeat.R
\name{neeat}
\alias{neeat}
\title{Network enhanced enrichment analysis tool}
\usage{
neeat(eval.gene.set, func.gene.sets, net, method = "neeat", max.depth = 1,
  rho = 0.5, n.perm = 1e+05, verbose = FALSE, n.cpu = 1,
  perm.batch = 10000)
}
\arguments{
\item{eval.gene.set}{Logical vector indicated the gene set for evaluating.}

\item{func.gene.sets}{Logical matrix indicated the functional gene sets associated with specific functions or pathways.
The rows correspond to genes, while the columns represent the functions or pathways.}

\item{net}{The adjacent matrix of network.}

\item{method}{A string indicated the NEEAT model, including "neeat", "neeat_hyper" and "hyper". 
Use "hyper" for traditional hypergeometric test, in which the network information is ignored.}

\item{max.depth}{Integer for the maximum depth considered in the NEEAT models.}

\item{rho}{The weight parameter for depths.}

\item{n.perm}{The number of permutations for calculating p-values. Minimum value 100 is required for the model "neeat".
Ignored in the models "neeat_hyper" and "hyper".}

\item{verbose}{Logical variable indicated whether output the full variables for computing p-value,
such as mean and variance.}

\item{n.cpu}{The number of CPUs/cores used in the parallel computation.}

\item{perm.batch}{The desired size of permutation batches in the parallel computation.}

\item{z.threshold}{The threshold for filtering out small Z-scores. The p-values are calculated only
for the Z-scores \code{>= z.threshold}, otherwise p-values are set as 1.}
}
\value{
This function will return a 2-dimensional array of dimensions \code{c(5, dim(func.core.sets)[2])},
and each column \code{[,j]} containing the following components for the correponding gene set for evaluating
and functional gene set \code{func.gene.sets[,j]}:
  \item{\code{z.score}}{The Z-score for the gene set}
  \item{\code{p.value}}{The statistic significance for the gene set under specified NEEAT model}
  \item{\code{raw.score}}{The raw score for the gene set under specified NEEAT model}
  \item{\code{avg.score}}{The average score for random permutations of the gene set}
  \item{\code{var.score}}{The variance of scores for random permutations of the gene set}
}
\description{
Functional enrichment anslysis of gene sets by integrating network information
}
\details{
This is a novel functional enrichment analysis tool based on the gene set and network
information.
}
\seealso{
\code{\link{get_func_gene_sets}}, \code{\link{neeat_depths}}
}

