% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depths.R
\name{neeat_depths}
\alias{neeat_depths}
\title{Compute the node depths}
\usage{
neeat_depths(seed.node.sets, net, max.depth)
}
\arguments{
\item{seed.node.sets}{Logical matrix indicated the seed nodes associated with specific conditions (gene sets).}

\item{net}{The adjacent matrix of network.}

\item{max.depth}{Integer for the maximum depth that will be computed.}
}
\value{
This function returns a matrix with the same dimensions as \code{seed.node.sets}, where the element
\code{[i, j]} represents the depth of node \code{i} under the condition (gene set) \code{j}.
}
\description{
Compute the node depths for given seed nodes and network
}
\details{
This function calculates the node depths, which is used in \code{\link{neeat}}.
}
\seealso{
\code{\link{neeat}}
}

