% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depths.R
\name{neeat_depths_with_permutation}
\alias{neeat_depths_with_permutation}
\title{Compute the node depths with permutation}
\usage{
neeat_depths_with_permutation(seed.nodes, net, n.perm, max.depth = 1,
  n.cpu = 1)
}
\arguments{
\item{seed.nodes}{Logical vector indicated the seed nodes.}

\item{net}{The adjacent matrix of network.}

\item{max.depth}{Integer for the maximum depth that will be computed.}

\item{n.cpu}{The number of CPUs/cores used in the parallel computation.}

\item{n.parm}{The number of permutations.}
}
\value{
This function returns a matrix with the dimensions \code{c(length(seed.nodes), n.perm+1)}, where the element
\code{[i, 1]} represents the depth of node \code{i}, and \code{[i, j+1]} the depth in permutation \code{j}.
}
\description{
Compute the node depths for the given seed node set and its permutations.
}
\details{
This function calculates the node depths of given seed nodes, togerther with the depths in random permutations of seed nodes.
}
\seealso{
\code{\link{neeat}}, \code{\link{neeat_depths}}
}

