% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgen.R
\name{netgen}
\alias{netgen}
\title{NetGen}
\usage{
netgen(annotation, PPI, active_gene, p1, p2, q, alpha = 3, trace = TRUE)
}
\arguments{
\item{annotation}{The input annotation matrix. Each row is a gene and each column is a functional category.}

\item{PPI}{The adjacent matrix of biological network.}

\item{active_gene}{The input character active gene list.}

\item{p1}{The probability of the core genes to be activated. 
If \code{p1} is a numeric value, \code{netgen} perform the enrichment analysis using the given parameter.
If \code{p1} is a numeric vector, \code{netgen} perform the enrichment analysis basing on a mixed parameter selection strategy.}

\item{p2}{The probability of the peripheral genes to be activated.
Both numeric value and numeric vector are acceptable as introduced in \code{p1}.}

\item{q}{The probability of all the other genes to be activated by external factor, such as noise, uncontrollable error in experiment, the incompletion of GO annotation and so on.
Both numeric value and numeric vector are acceptable as introduced in \code{p1}.}

\item{alpha}{A positive number to balance the log-likelihood and penalization term.
Pre-set value is 3.}

\item{trace}{Logical variable indicated whether tracing information on the solving progress is produced.}
}
\value{
The returns of this function \code{fresult} depend on the model input parameter \code{p1}, \code{p2} and \code{q}. 
A matrix of enriched categories and its corresponding Fisher's exact test p-value is returned if the the model input parameter \code{p1}, \code{p2} and \code{q} are all numeric values.
A list is returned if at least one of the model input parameter \code{p1}, \code{p2} and \code{q} is numeric vector.
Each element of the list is a parameter combination result matrix and a combined p-value.
}
\description{
A novel enhanced network-based generative model for gene set functional enrichment analysis
}
\details{
This function perform the gene set enrichment analysis using network-based generative model, NetGen. 
An additional protein-protein interaction (PPI) network was explicitly used to assist the functional analysis.
A greedy-based approximate algorithm was peformed to seek for a sub-optimal solution of the log-likelihood function.
}
\references{
Duanchen Sun, Yinliang Liu, Xiang-Sun Zhang, Ling-Yun Wu.
NetGen: a novel network-based probabilistic generative model for gene set functional enrichment analysis.
BMC Systems Biology, 11(Suppl 4):75, 2017.
}
