% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotation.R
\name{require_bioc}
\alias{require_bioc}
\title{Load a Bioconductor package}
\usage{
require_bioc(package)
}
\arguments{
\item{package}{The name of package name.}
}
\description{
Load a Bioconductor package, automatically install if need.
}

