% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{submatrix}
\alias{submatrix}
\title{Extract a sub-matrix from a matrix}
\usage{
submatrix(m, rows, cols)
}
\arguments{
\item{m}{The matrix}

\item{rows}{The row indexes}

\item{cols}{The column indexes}
}
\value{
This function will return the specified sub-matrix as a matrix of corresponding type.
}
\description{
Extract a specified sub-matrix from a sparse matrix rapidly
}
\details{
This function use faster extraction algorithm for the \code{\link[=CsparseMatrix-class]{CsparseMatrix}} class in the package \pkg{Matrix}.
}
