% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_significant_terms}
\alias{get_significant_terms}
\title{Get significant term list}
\usage{
get_significant_terms(p.values, term.info, threshold = 0.05)
}
\arguments{
\item{p.values}{The vector of p-values.}

\item{term.info}{The matrix of term information with three columns: Term ID, Term category,
 and Term description. The row names should be set as Term ID.
 See \code{\link{get_annotation}} for more information.}
}
\value{
This function returns a matrix with four columns: Rank, Term ID, Term category,
 and Term description.
}
\description{
Extract and rank the significant terms with p-values smaller than the given threshold.
}
\seealso{
\code{\link{get_annotation}}
}

