% Generated by roxygen2 (4.0.1): do not edit by hand
\name{id_mapping_species}
\alias{id_mapping_species}
\title{Mapping gene central id from one species to other species}
\usage{
id_mapping_species(ids, species.from, species.to)
}
\description{
Mapping gene central id from one species to other species
}

