% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neeat.R
\name{neeat}
\alias{neeat}
\title{Network enhanced enrichment analysis tool}
\usage{
neeat(eval.gene.set, func.gene.sets, net, method = "neeat", max.depth = 1,
  rho = 1, n.perm = 1e+05, keep.degree = TRUE, n.cpu = 1,
  perm.batch = 5000)
}
\arguments{
\item{eval.gene.set}{Logical vector indicated the gene set for evaluating.}

\item{func.gene.sets}{Logical matrix indicated the functional gene sets associated with specific functions or pathways.
The rows correspond to genes, while the columns represent the functions or pathways.}

\item{net}{The adjacent matrix of network.}

\item{method}{A string indicated the NEEAT model, including "neeat", "neeat_h1", "neeat_h2" and "hyper". 
Use "hyper" for traditional hypergeometric test, in which the network information is ignored.}

\item{max.depth}{Integer for the maximum depth considered in the NEEAT models.}

\item{rho}{The weight parameter for depths.}

\item{n.perm}{The number of permutations for calculating p-values. Minimum value 100 is required for the model "neeat".
Ignored in the models "neeat_h1", "neeat_h2" and "hyper".}

\item{keep.degree}{Logical variable indicated whether keep the degree distribution as much as possible in permutations.}

\item{n.cpu}{The number of CPUs/cores used in the parallel computation.}

\item{perm.batch}{The desired number of permutations in each batch of the parallel computation.}
}
\value{
This function will return a 2-dimensional array of dimensions \code{c(5, dim(func.gene.sets)[2])},
and each column \code{[,j]} containing the following components for the corresponding functional gene set:
  \item{\code{z.score}}{The Z-score of the functional gene set}
  \item{\code{p.value}}{The statistic significance of the functional gene set}
  \item{\code{avg.score}}{The average score in random permutations}
  \item{\code{var.score}}{The variance of scores in random permutations}
  \item{\code{raw.score}}{The raw score of the functional gene set}
}
\description{
Functional enrichment anslysis of gene sets by integrating network information
}
\details{
This is a novel functional enrichment analysis tool based on the gene set and network
information.
}
\seealso{
\code{\link{get_func_gene_sets}}, \code{\link{neeat_depths}}
}

