% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{require_bioc}
\alias{require_bioc}
\title{Load a Bioconductor package}
\usage{
require_bioc(pkg.name)
}
\arguments{
\item{pkg.name}{The name of package.}
}
\description{
Load a Bioconductor package, automatically install if need.
}
